requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
tensorflow==2.13
deepface==0.0.79
Pillow>=9.0
streamlit>=1.25
tabulate>=0.9
